[arren] vCircle updated with hough transform methods - modularisation of houghFlow and window size (lifetime), i.e. these can be separated if needed - reduced hough mapping to a 3x3 region - circle sizes changed to 10->48 pixels - inlierThreshold now a percentage of the circle visible - better stat dump handling from command line etc.

+++ src/processing/circle/include/vCircleModule.h
@@ -45,7 +45,7 @@ public:
     vGeoCircleObserver geomFinder;
     vHoughCircleObserver houghFinder;
     vCircleTracker circleTracker;
-    int inlierThreshold;
+    double inlierThreshold;
     bool hough;
     
     vCircleReader();
